Add DataModel to handle data related tasks

+++ origami/Model.py
@@ -4,59 +4,36 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 import warnings
 import shutil
-import os
-from typing import Tuple, final, Any
-from pathlib import Path
-import joblib  # type: ignore
-from joblib import Memory
-import numpy as np
+from typing import final, Any
+from joblib import Memory # type: ignore
 
 from .Serializable import Serializable
+from .DataModel import DataModel, IntegrityError
 
+class ModelNotTrained(UserWarning):
+    """ Raised when trying to use an untrained model to predict """
 
 class Model(Serializable, ABC):
     """ A base class for any predictive model """
     trained: bool = False
-    input_shape: Tuple[int, ...]
-    output_shape: Tuple[int, ...]
     _cache_memory: Memory
 
-    def __init__(self, input_shape: Tuple[int, ...], output_shape: Tuple[int, ...], *args, **kwargs):
-        """
-        Initialises the model object
-
-        :param input_shape: constraint on the last dimensions of input data (for both training and predicting)
-        :param output_shape: constraint on the last dimensions of the output data (for predicting only)
-
-        The input shape should match:
-
-        >>> model = Model(input_shape=(10,3), output_shape=(42,))
-        >>> model.train(np.random.random((300, 10, 4)))
-            AssertionError: Wrong shape
-        >>> model.train(np.random.random((300, 10, 3)))
-        >>> model.predict(np.random.random((123, 10, 4)))
-            AssertionError: Wrong shape
-
-        And it's garanteed that the output shape will match too:
-
-        >>> pred = model.predict(np.random.random((123, 10, 3)))
-        >>> assert pred.shape[-1] == 42
-        """
-        self.input_shape = input_shape
-        self.output_shape = output_shape
+    def __init__(self, *args, **kwargs):
+        """ Initialises the model object """
         self._cache_memory = Memory(location=None, verbose=False)
 
     @final
-    def train(self, data: np.ndarray, *args: Any, **kwargs: Any) -> None:
+    def train(self, data: DataModel, /, *args: Any, **kwargs: Any) -> None:
         """
         Trains the model with the provided data and arguments.
 
-        :param data: Data with which to train, should have the same last dimensions
-         as the `input_shape`
+        :param data: Data with which to train
         """
-        assert isinstance(data, np.ndarray)
-        # Forces the input to fit in the given shape
-        assert data.shape[-len(self.input_shape):] == self.input_shape, "Wrong shape"
+        if not isinstance(data, DataModel):
+            raise TypeError("Argument data should be of type DataModel,"\
+                           f"found {data.__class__.__name__}")
+        if not data.check_integrity():
+            raise IntegrityError("Data failed integrity check before training")
         self._train(data, *args, **kwargs)
         self.trained = True
         # Note: Cache isn't shared between "freshly created" instances and "loaded from file" instances !
@@ -64,35 +41,42 @@ class Model(Serializable, ABC):
         self._cache_memory = Memory(location=f"./model_cache", verbose=False)
 
     @final
-    def predict(self, data: np.ndarray, *args: Any, **kwargs: Any) -> np.ndarray:
+    def predict(self, data: DataModel, /, *args: Any, check_output=True, **kwargs: Any) -> DataModel:
         """ Predict data from the model,
 
-        :param data: Data from which to predict, should have the same last dimensions
-         as the `input_shape`
-        :return: Predicted data, will have the same last dimensions
-         as the `output_shape`
+        :param data: Data from which to predict
+        :param check_output: Whether to check the prediction's output integrity.
+         May be frustrating when debugging, you may want to turn this off then.
+        :return: Predicted data
         """
-        assert isinstance(data, np.ndarray) and self.trained
-        # Forces the input to fit in the given shape
-        assert data.shape[-len(self.input_shape):] == self.input_shape, "Wrong shape"
+        if not self.trained:
+            raise ModelNotTrained("Trying to predict using an untrained model !"\
+                                  "Results may not be satisfying.")
+        if not isinstance(data, DataModel):
+            raise TypeError("Argument data should be of type DataModel,"\
+                           f"found {data.__class__.__name__}")
+        if not data.check_integrity():
+            raise IntegrityError("Data failed integrity check before predicting")
+
         # Tries to use self._predict's cached results first
         predict_cached = self._cache_memory.cache(self._predict)
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
             output = predict_cached(data, *args, **kwargs)
-        # Forces the output to fit in the given shape
-        assert output.shape[-len(self.output_shape):] == self.output_shape, "Wrong shape"
+        if check_output and not output.check_integrity():
+            raise IntegrityError("Prediction failed integrity check")
         return output
 
-    def clear_cache(self):
+    def clear_cache(self, *, soft=False):
         """ Clears the model's prediction cache """
         self._cache_memory.clear(warn=False)
-        shutil.rmtree(self._cache_memory.location, ignore_errors=True)
+        if not soft:
+            shutil.rmtree(self._cache_memory.location, ignore_errors=True)
 
     @abstractmethod
-    def _train(self, data: np.ndarray, *args: Any, **kwargs: Any) -> None:
+    def _train(self, data: DataModel, *args: Any, **kwargs: Any) -> None:
         raise NotImplementedError()
 
     @abstractmethod
-    def _predict(self, data: np.ndarray, *args: Any, **kwargs: Any) -> np.ndarray:
+    def _predict(self, data: DataModel, *args: Any, **kwargs: Any) -> DataModel:
         raise NotImplementedError()
